Remove duplicated code in convention plugins
